Creating myBankPackage (started v0.1.1)
