+++ .env
@@ -1,0 +1,2 @@
+BUDGET_PATH="../storage/budgets.csv"
+ACOUNT_PATH="../storage/accounts.json"
