+++ .gitignore
@@ -1,0 +1,12 @@
+# Ignore pycache
+__pycache__/
+
+# Ignore pytest cache
+.pytest_cache/
+
+# Ignore .venv
+.venv/
+
+# Ignore build and egg-info directories
+build/
+*.egg-info/
