+++ myBankPackage/__init__.py
@@ -1,0 +1,190 @@
+__version__ = '0.1.1'
+
+
+
+import uuid
+import os
+import json
+import pandas as pd
+from datetime import datetime
+from dataclasses import dataclass, field
+
+
+"""
+PATHS
+"""
+# Path to the budget CSV file
+budget_path = os.getenv("BUDGET_PATH", "../storage/test_budgets.csv") 
+
+# Path to the account JSON file
+account_path = os.getenv("ACCOUNT_PATH", "../storage/test_accounts.json")
+
+"""
+VALUE RESTRICTIONS
+"""
+available_account_types = ("checking", "savings")
+
+
+
+
+
+### COMMON FUNCTIONS ###
+def generate_uuid():
+    """
+    Generates a random UUID used for acocunt ID, transaction ID, etc.
+    """
+    return uuid.uuid4().hex
+
+
+
+"""
+ACCOUNT
+"""
+
+### Account Dataclass ###
+@dataclass(frozen=False, order=False)
+class Account:
+    """
+    Represents a bank account.
+
+    Attributes:
+        name (str): The name of the account.
+        type (str): The type of the account (checking or savings).
+        amount (float): The amount of money in the account.
+        id (str): The unique identifier of the account.
+        history (list): The transaction history of the account.
+        _search_string (str): A string used for searching the account.
+
+    Methods:
+        __post_init__(): Initializes the account object.
+        deposit(amount: float, transaction_id): Deposits money into the account.
+        withdraw(amount: float, transaction_id): Withdraws money from the account.
+        save(account_path: str = account_path): Saves the account to a JSON file without overwriting existing data.
+    """
+    # Name of the account
+    name:str
+    # Type of the account (checking or savings)
+    type:str = "checking"
+    # Amount of the account
+    amount:float = 0.0
+
+    ### NO INIT ###
+
+    # Random ID is generated by a function
+    id: str = field(init=False, default_factory=generate_uuid)
+    # History
+    history: list = field(init=False, default_factory=list)
+
+
+    def __post_init__(self) -> None:
+        # Check if the type is an instance of available_account_types
+        if self.type not in available_account_types:
+            raise TypeError(f"Invalid Account Type: {self.type}")
+        # Check if the amount is a float and positive
+        if self.amount < 0:
+            raise ValueError(f"Invalid Account Amount: {self.amount}")
+        
+        print(f"Successfully created account ! \n" \
+                              f"Name: {self.name} \n" \
+                              f"Type: {self.type} \n" \
+                              f"Amount: {self.amount}")
+
+
+    def deposit(self, amount:float, transaction_id) -> None:
+        """
+        Deposit money into the account
+
+        Args:
+            amount (float): The amount of money to deposit.
+            transaction_id: The ID of the transaction.
+
+        Raises:
+            ValueError: If the amount is negative.
+
+        Returns:
+            None
+        """
+        if amount < 0:
+            raise ValueError(f"Invalid Deposit Amount: {amount}")
+        self.amount += amount
+
+        self.history.append({
+            "date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+            "transaction_id": transaction_id,
+            "type": "deposit",
+            "amount": amount,
+        })
+
+        print(f"Successfully deposited {amount} into account {self.name}. \n New amount: {self.amount}")
+
+
+    def withdraw(self, amount:float, transaction_id) -> None:
+        """
+        Withdraw money from the account
+
+        Args:
+            amount (float): The amount of money to withdraw.
+            transaction_id: The ID of the transaction.
+
+        Raises:
+            ValueError: If the amount is negative or exceeds the account balance.
+
+        Returns:
+            None
+        """
+        if amount < 0:
+            raise ValueError(f"Invalid Withdraw Amount: {amount}")
+        if amount > self.amount:
+            raise ValueError(f"Insufficient funds to withdraw {amount} from account {self.name}.")
+        
+        
+        self.amount -= amount
+
+        self.history.append({
+            "date": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
+            "transaction_id": transaction_id,
+            "type": "withdraw",
+            "amount": amount,
+        })
+
+        print(f"Successfully withdrew {amount} from account {self.name}. \n New amount: {self.amount}")
+
+
+    def save(self, account_path:str=account_path) -> None:
+        """
+        Save the account to a JSON file without overwriting existing data
+
+        Args:
+            account_path (str): The path to the JSON file.
+
+        Returns:
+            None
+        """
+        try:
+            with open(account_path, "r") as file:
+                data = json.load(file)
+        except FileNotFoundError:
+            data = {}
+
+        # Update the data dictionary to include the new account
+        data[self.name] = self.__dict__
+
+        with open(account_path, "w") as file:
+            json.dump(data, file, indent=4)
+
+
+
+### Account Functions ###
+def create_account(name: str, type:str, amount: float = 0.0):
+    """
+    Creates a new account object.
+
+    Parameters:
+        name (str): The name of the account.
+        account_type (str): The type of the account (checking or savings).
+        amount (float): The initial amount of the account.
+
+    Returns:
+        Account: The newly created account object.
+    """
+    return Account(name, type, amount)
