+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "my-bank-app"
+version = "0.1.1"
+description = ""
+authors = ["Guillaume pot <guillaume.pot@outlook.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.10"
+pytest = "^8.2.0"
+pandas = "^2.2.2"
+streamlit = "^1.34.0"
+
+
+[tool.poetry.group.dev.dependencies]
+ipykernel = "^6.29.4"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
