--- src/.env
@@ -1,2 +1,0 @@
-BUDGET_PATH="../storage/test_budgets.csv"
-ACOUNT_PATH="../storage/test_accounts.json"
