+++ storage/test_accounts.json
@@ -1,44 +1,2 @@
 {
-    "test 1": {
-        "name": "test 1",
-        "type": "checking",
-        "amount": -10.5,
-        "id": "b831b1e401e54ea48c76029ec8a615d2",
-        "history": [
-            {
-                "date": "2024-05-04 19:19:04",
-                "transaction_id": "1",
-                "type": "deposit",
-                "amount": 100
-            },
-            {
-                "date": "2024-05-04 19:19:04",
-                "transaction_id": "2",
-                "type": "withdraw",
-                "amount": 1050.5
-            }
-        ],
-        "_search_string": "Successfully created account ! \nName: test 1 \nType: checking \nAmount: 1000"
-    },
-    "test 2": {
-        "name": "test 2",
-        "type": "savings",
-        "amount": 450,
-        "id": "d3449175d48546d9be66694520337a95",
-        "history": [
-            {
-                "date": "2024-05-04 19:19:04",
-                "transaction_id": "1",
-                "type": "deposit",
-                "amount": 50
-            },
-            {
-                "date": "2024-05-04 19:19:04",
-                "transaction_id": "2",
-                "type": "withdraw",
-                "amount": 100
-            }
-        ],
-        "_search_string": "Successfully created account ! \nName: test 2 \nType: savings \nAmount: 500"
-    }
 }
