+++ storage/test_budgets.csv
@@ -1,2 +1,1 @@
 id,name,month,amount,history
-bef36af93cac4009b2c5248f839468a2,test,August,550,"[{'date': '2024-05-04 19:05:40', 'transaction_id': '2', 'amount': 50}]"
